[Nette] Rename AddDatePickerToDateControlRector to AddNextrasDatePickerToDateControlRector
