+++ config/set/nette-30.php
@@ -7,7 +7,7 @@ use Rector\Generic\Rector\ClassMethod\ArgumentDefaultValueReplacerRector;
 use Rector\Generic\Rector\MethodCall\FormerNullableArgumentToScalarTypedRector;
 use Rector\Generic\ValueObject\ArgumentDefaultValueReplacer;
 use Rector\Nette\Rector\ClassMethod\RemoveParentAndNameFromComponentConstructorRector;
-use Rector\Nette\Rector\MethodCall\AddDatePickerToDateControlRector;
+use Rector\Nette\Rector\MethodCall\AddNextrasDatePickerToDateControlRector;
 use Rector\Nette\Rector\MethodCall\GetConfigWithDefaultsArgumentToArrayMergeInCompilerExtensionRector;
 use Rector\Nette\Rector\MethodCall\MagicHtmlCallToAppendAttributeRector;
 use Rector\Nette\Rector\MethodCall\RequestGetCookieDefaultArgumentToCoalesceRector;
@@ -27,7 +27,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $containerConfigurator->import(__DIR__ . '/nette-30-return-types.php');
     $containerConfigurator->import(__DIR__ . '/nette-30-param-types.php');
     $services = $containerConfigurator->services();
-    $services->set(AddDatePickerToDateControlRector::class);
+    $services->set(AddNextrasDatePickerToDateControlRector::class);
     $services->set(ChangeFormArrayAccessToAnnotatedControlVariableRector::class);
     $services->set(GetConfigWithDefaultsArgumentToArrayMergeInCompilerExtensionRector::class);
     // Control class has remove __construct(), e.g. https://github.com/Pixidos/GPWebPay/pull/16/files#diff-fdc8251950f85c5467c63c249df05786
