+++ rules/nette/tests/Rector/MethodCall/AddNextrasDatePickerToDateControlRector/Fixture/fixture.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Nette\Tests\Rector\MethodCall\AddDatePickerToDateControlRector\Fixture;
+namespace Rector\Nette\Tests\Rector\MethodCall\AddNextrasDatePickerToDateControlRector\Fixture;
 
 use Nette\Application\UI\Form;
 
@@ -17,7 +17,7 @@ class SomeClass
 -----
 <?php
 
-namespace Rector\Nette\Tests\Rector\MethodCall\AddDatePickerToDateControlRector\Fixture;
+namespace Rector\Nette\Tests\Rector\MethodCall\AddNextrasDatePickerToDateControlRector\Fixture;
 
 use Nette\Application\UI\Form;
